chore(ci): enforce C++ skill standards (#64)
* chore(ci): enforce C++ skill standards

* fix(ci): avoid GCC shadow diagnostics

* fix(tests): avoid Qt timeout conversion warnings

+++ Tests/Unit/Engine/UnitTest_Replay.cpp
@@ -37,7 +37,7 @@ public:
     if (consumed_ >= static_cast<std::int64_t>(bars_.size())) {
       return std::nullopt;
     }
-    return bars_[consumed_++];
+    return bars_[static_cast<std::size_t>(consumed_++)];
   }
 
   std::int64_t totalBars() const noexcept override {
